Update post.go
Hi, first ok all compliment for your amazing contents. Try to understanding better the refactor you did on productAPI (out-screen between lesson 6 and 7) I find this bug which compromise both the post and put requests.
The error is the following:
"interface {} is nil, not validator.ValidationErrors"
And this is due to the fact that row 20 extract from the context a pointer to Product witch can not be cast to Product.
Hope you'll find useful this reporting. Keep going with your video you are an amazing teacher! =)